# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/films/halls/hall": {
            "get": {
                "tags": [
                    "Films",
                    "Halls",
                    "Unfinished"
                ],
                "summary": "Get Halls",
                "operationId": "get_halls_films_halls_hall_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "read:halls"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "hall_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Hall Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hall"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Films",
                    "Halls",
                    "Unfinished"
                ],
                "summary": "Create Hall",
                "operationId": "create_hall_films_halls_hall_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "write:halls"
                        ]
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Hall"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hall"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Films",
                    "Halls",
                    "Unfinished"
                ],
                "summary": "Update Hall",
                "operationId": "update_hall_films_halls_hall_patch",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "write:halls"
                        ]
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Hall"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/films/halls/hall/id/{hall_id}": {
            "get": {
                "tags": [
                    "Films",
                    "Halls",
                    "Unfinished"
                ],
                "summary": "Get Halls By Hall Id",
                "operationId": "get_halls_by_hall_id_films_halls_hall_id__hall_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "read:halls"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "hall_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Hall Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hall"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/films/halls/halls": {
            "get": {
                "tags": [
                    "Films",
                    "Halls",
                    "Unfinished"
                ],
                "summary": "Get Halls",
                "operationId": "get_halls_films_halls_halls_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "read:halls"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "title": "Start id"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 2,
                            "title": "limit id to update"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Hall"
                                    },
                                    "title": "Response Get Halls Films Halls Halls Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/films/halls/hall/schedules/{hall_id}/{limit}": {
            "get": {
                "tags": [
                    "Films",
                    "Halls",
                    "Unfinished"
                ],
                "summary": "Get Schedules",
                "operationId": "get_schedules_films_halls_hall_schedules__hall_id___limit__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": [
                            "read:halls"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "name": "hall_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Hall Id"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Hall": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "hall_name": {
                        "type": "string",
                        "title": "Hall Name"
                    },
                    "seats_per_row": {
                        "type": "integer",
                        "minimum": 1.0,
                        "title": "Seats Per Row"
                    },
                    "no_of_rows": {
                        "type": "integer",
                        "minimum": 1.0,
                        "title": "No Of Rows"
                    },
                    "class_name": {
                        "type": "string",
                        "title": "Class Name",
                        "default": "HALL"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "hall_name",
                    "seats_per_row",
                    "no_of_rows"
                ],
                "title": "Hall"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "token"
                    }
                }
            }
        }
    }
}